Core: Fix warnings and improve formatting

+++ core/src/dlg_imageproperty.h
@@ -33,7 +33,7 @@ class dlg_imageproperty : public QDockWidget, public Ui_ImageProperty
 
 public:
 	dlg_imageproperty(QWidget *parent);
-	void AddInfo(vtkImageData* src, iAImageInfo const & info, QString const & name, int channelCount);
+	void AddInfo(vtkImageData* src, iAImageInfo const & info, QString const & name, size_t channelCount);
 	void Clear();
 private:
 	void EnterMsg(QString txt);
